requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2

plotly==5.24.1
streamlit==1.39.0
requests==2.32.3
transformers
torch

